pyproject.toml
[project]
name = "trend-to-rule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.12"
dependencies = [
    "feedparser>=6.0.12",
    "flagembedding>=1.3.5",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "lmdb>=1.7.5",
    "openai>=2.26.0",
    "orjson>=3.11.7",
    "pydantic>=2.11.7",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "streamlit>=1.55.0",
    "transformers==4.57.6",
]
